Chore: TagEntity 파일 분리 #10
